chore: move group chat logic to separate file
